+++ crates/dxm-init/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   hard-coding paths.
 - **Breaking:** updated `server` to use paths from the manifest instead of
   hard-coded paths.
+- Updated `vcs::VcsOption` to derive from `Debug`, `PartialEq` and `Eq`.
 
 ## 0.1.2 - 2026-01-12